requirements.txt
requests
xlsxwriter
aiohttp
tqdm
pydantic
pydantic-settings
urllib3
black
flake8
pytest
isort
PyYAML
backoff
